Moved pages into views folder, created component for RandomPost page, added comments to component
