+++ src/views/RandomBoard.vue
